# mobile/backend/docs/dating-api-spec.yaml
openapi: "3.0.3"
info:
  title: Dating API
  version: "1.0.0"
  description: |
    REST API для dating-приложения. Монорепо: backend (Go/chi) + frontend (React SPA).

    **Формат всех ответов:**
    ```json
    { "data": <payload | null>, "error": <"сообщение" | null> }
    ```

    **Rate limits:**
    - `/api/auth/*` — 5 rps, burst 10 (per IP)
    - `/api/*` (защищённые) — 100 rps, burst 200 (per IP)

    **Body limit:** 1 МБ на все запросы.
servers:
  - url: http://localhost:8080
    description: Локальная разработка
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access-токен, выданный при регистрации, логине или OAuth.
        TTL по умолчанию: 15 минут.
  schemas:
...
    Envelope:
      type: object
      required: [data, error]
      properties:
        data:
          description: Полезная нагрузка (null при ошибке)
        error:
          type: string
          nullable: true
          description: Сообщение об ошибке (null при успехе)

    # ── Пользователь ──────────────────────────────────────────────────────────
    ProfileResponse:
      type: object
      description: Публичный профиль пользователя, возвращаемый API.
      required: [id, name, age, bio, interests, location, photo, birthDate]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: display_name из БД
          example: "Алексей"
        age:
          type: integer
          description: Вычисляется из birth_date; 0 если дата не задана
          example: 28
        bio:
          type: string
          description: Описание профиля (пустая строка если не задано)
          example: "Люблю горы и кофе"
        interests:
          type: array
          items:
            type: string
          description: Список интересов (JSONB → []string)
          example: ["hiking", "coffee", "dev"]
        location:
          type: string
          description: Город/локация (пустая строка если не задано)
          example: "Москва"
        photo:
          type: string
          description: avatar_url (пустая строка если не задано)
          example: "https://cdn.example.com/media/abc.jpg"
        birthDate:
          type: string
          format: date
          description: Дата рождения в формате YYYY-MM-DD (пустая строка если не задана)
          example: "1996-03-15"
        personality:
          description: Произвольный JSONB-объект данных личности (опционально)
          nullable: true
          example: { "type": "INTJ", "traits": ["analytic"] }
        astrology:
          description: Произвольный JSONB-объект астрологических данных (опционально)
          nullable: true
          example: { "sign": "Рыбы", "element": "Вода" }
        numerology:
          description: Произвольный JSONB-объект нумерологических данных (опционально)
          nullable: true
          example: { "lifePathNumber": 7 }
    AuthResponse:
      type: object
      description: Ответ на успешную аутентификацию (register / login / OAuth).
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
          description: JWT HS256 access-токен, TTL 15 минут
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Opaque refresh-токен, TTL 30 дней (720 ч)
          example: "dGhpcyBpcyBhIHJhbmRvbSByZWZyZXNoIHRva2Vu"
        user:
          $ref: "#/components/schemas/ProfileResponse"
    TokenPair:
      type: object
      description: Пара токенов без данных пользователя (для /auth/refresh).
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "dGhpcyBpcyBhIHJhbmRvbSByZWZyZXNoIHRva2Vu"

    # ── Сообщения ─────────────────────────────────────────────────────────────
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      description: Оболочка при ошибке.
      required: [data, error]
      properties:
        data:
          nullable: true
          example: null
        error:
          type: string
          example: "неверный email или пароль"
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      description: |
        Создаёт учётную запись с email+password. При успехе возвращает
        access + refresh токены и профиль. Пароль минимум 6 символов.
        При дублировании email — 409 Conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, birthDate, email, password]
              properties:
                name:
                  type: string
                  description: Отображаемое имя (display_name)
                  example: "Алексей"
                birthDate:
                  type: string
                  format: date
                  description: Дата рождения YYYY-MM-DD
                  example: "1996-03-15"
                email:
                  type: string
                  format: email
                  example: "alex@example.com"
                phone:
                  type: string
                  description: Номер телефона (опционально)
                  example: "+79991234567"
                password:
                  type: string
                  minLength: 6
                  example: "s3cr3tP@ss"
                agreeToPrivacy:
                  type: boolean
                  example: true
                agreeToTerms:
                  type: boolean
                  example: true
                agreeToOffer:
                  type: boolean
                  example: true
      responses:
        "201":
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Неверный формат запроса или пароль короче 6 символов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email уже занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Вход по email и паролю
      description: |
        Проверяет email+password и выдаёт access + refresh токены.
        Email приводится к нижнему регистру и обрезается.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "alex@example.com"
                password:
                  type: string
                  example: "s3cr3tP@ss"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление пары токенов
      description: |
        Принимает действующий refresh-токен и возвращает новую пару
        access + refresh (rotation). Старый refresh-токен аннулируется.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "dGhpcyBpcyBhIHJhbmRvbSByZWZyZXNoIHRva2Vu"
      responses:
        "200":
          description: Новая пара токенов
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TokenPair"
        "400":
          description: refreshToken отсутствует или неверный формат запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Токен недействителен или истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Выход (отзыв refresh-токена)
      description: |
        Аннулирует переданный refresh-токен. Ошибки DB игнорируются —
        если токен уже отозван, ответ всё равно 200.
        Клиент должен самостоятельно удалить access-токен из памяти.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "dGhpcyBpcyBhIHJhbmRvbSByZWZyZXNoIHRva2Vu"
      responses:
        "200":
          description: Токен отозван (или уже был отозван)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OkResponse"
        "400":
          description: refreshToken отсутствует или неверный формат запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/yandex/callback:
    get:
      tags: [Auth]
      summary: Обработка callback от Yandex OAuth
      description: |
        Принимает `code` и `state` от Yandex, обменивает code на токен,
        получает профиль через `https://login.yandex.ru/info?format=json`,
        создаёт или обновляет пользователя (upsert по yandex_id) и
        возвращает пару токенов + профиль.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code от Yandex
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF state (должен совпадать с выданным в /api/auth/yandex)
      responses:
        "200":
          description: Успешная аутентификация через Yandex
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Неверный/истёкший state или ошибка обмена кода
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Ошибка при обращении к Yandex API за профилем
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/messenger/callback:
    get:
      tags: [Auth]
      summary: Обработка callback от MAX Messenger OAuth (заглушка)
      description: |
        **STUB — не реализовано.** Всегда возвращает 501.
        TODO(MESSENGER_PROVIDER): реализовать после получения credentials от MAX Messenger.
      parameters:
        - name: code
          in: query
          schema:
            type: string
          description: Authorization code от MAX Messenger (не обрабатывается)
        - name: state
          in: query
          schema:
            type: string
          description: CSRF state (не обрабатывается)
      responses:
        "501":
          description: Провайдер не настроен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                data: null
                error: "MESSENGER_PROVIDER не настроен"

  # ════════════════════════════════════════════════════════════════════════════
  # USERS
  # ════════════════════════════════════════════════════════════════════════════
...